+++ RealWorld/world/physics/Hitbox.hpp
@@ -3,22 +3,22 @@
 
 class Hitbox {
 public:
-	Hitbox(const glm::vec2& botLeftPx, const glm::vec2& dimsPx, const glm::vec2& centerOffsetPx = glm::vec2(0.0f));
+	Hitbox(const glm::ivec2& botLeftPx, const glm::ivec2& dimsPx, const glm::ivec2& centerOffsetPx = glm::ivec2(0.0f));
 	~Hitbox();
 
-	glm::vec2& botLeft();
-	const glm::vec2& getBotLeft() const;
+	glm::ivec2& botLeft();
+	const glm::ivec2& getBotLeft() const;
 
-	void setDims(const glm::vec2& dimensionPx);
+	void setDims(const glm::ivec2& dimensionPx);
 
-	glm::vec2 getCenter() const;
+	glm::ivec2 getCenter() const;
 	bool collidesWith(Hitbox& otherHitbox) const;
-	bool overlaps(const glm::vec2& pointPx) const;
+	bool overlaps(const glm::ivec2& pointPx) const;
 
-	bool overlapsBlockwise(const glm::vec2& pointPx) const;
+	bool overlapsBlockwise(const glm::ivec2& pointPx) const;
 protected:
 	//Basic physics variables
-	glm::vec2 p_botLeftPx;//Bottom-left corner of the hitbox
-	glm::vec2 p_dimsPx;
-	glm::vec2 p_centerOffsetPx;//m_botLeftPx + m_offsetPx -> considered as center of the hitbox
+	glm::ivec2 p_botLeftPx;//Bottom-left corner of the hitbox
+	glm::ivec2 p_dimsPx;
+	glm::ivec2 p_centerOffsetPx;//m_botLeftPx + m_offsetPx -> considered as center of the hitbox
 };
